+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 = v7.x.x - 2025-x-x =
 - **Enhancement:** Refactored the FARAZSMS gateway for improved reliability.
+- New: Support SMSALA gateway in WP SMS Pro.
 
 = v7.0.3 - 2025-09-17 =
 - **Enhancement:** Improved Send SMS page performance by loading recipients via AJAX instead of on initial render.